+++ backends-velox/src-iceberg/main/scala/org/apache/gluten/extension/IcebergPostTransform.scala
@@ -14,8 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.gluten.execution
+package org.apache.gluten.extension
 
+import org.apache.gluten.execution.{VeloxIcebergAppendDataExec, VeloxIcebergOverwriteByExpressionExec, VeloxIcebergReplaceDataExec}
 import org.apache.gluten.extension.injector.Injector
 
 import org.apache.spark.sql.catalyst.rules.Rule
